der: extract `length::indefinite` module
Extracts a module containing everything related to indefinite length
handling in a single place
